requirements.txt
streamlit
pandas
openpyxl
chromadb
google-generativeai
